chore(linux): Add build.sh to ibus-keyman 🏗️

+++ core/build.sh
@@ -158,7 +158,7 @@ displayInfo "" \
     ""
 
 clean() {
-  rm -rf "$TARGET_PATH/"
+  rm -rf "${TARGET_PATH:?}/"
 }
 
 path_remove() {
